requirements.txt
numpy==2.1.3
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
pillow==11.3.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
dlib==20.0.0
mediapipe==0.10.14
deepface==0.0.93
transformers==4.53.3
torch==2.7.1
sounddevice==0.5.2
simpleaudio==1.0.4
pygame==2.6.1
pydub==0.25.1
edge-tts==7.0.2
requests==2.32.4
